Add files via upload
This is a simple script in order to identify and visualise fires detected using Sentinel 2 imagery. The script requires validation and should maybe be used in training a machine lkearning algorithm in order to improve accuracy. there is a slight problem caused by differences in pixel size between the SWIR bands and the visual spectrum bands which once corrected for in the script will greatly reduce the number of false positives